refactor common logic - re-export refs
