+++ tools/apiview/parsers/rust-api-parser/src/process-items/utils/typeToReviewTokens.ts
@@ -1,8 +1,6 @@
 import { Type } from "../../../rustdoc-types/output/rustdoc-types";
-import { getAPIJson } from "../../main";
 import { ReviewToken, TokenKind } from "../../models/apiview-models";
-import { reexportLines } from "../processUse";
-import { externalReexports } from "./externalReexports";
+import { addExternalReexportIfNotExists } from "./externalReexports";
 import { processFunctionPointer } from "./processFunctionPointer";
 import { processGenericArgs } from "./processGenerics";
 import { shouldElideLifetime } from "./shouldElideLifeTime";
@@ -23,13 +21,8 @@ export function typeToReviewTokens(type: Type): ReviewToken[] {
       NavigateToId: type.resolved_path.id.toString(),
     };
 
-    // If the reexportLines.external.items does not already contain the line, add it
-    if (
-      !reexportLines.external.items.some((line) => line.LineId === type.resolved_path.id.toString())
-    ) {
-      const lines = externalReexports(type.resolved_path.id, getAPIJson());
-      reexportLines.external.items.push(...lines);
-    }
+    // Add external re-export if it's an external type and not already added
+    addExternalReexportIfNotExists(type.resolved_path.id);
     // If there are no generic arguments, just return the base token
     if (!type.resolved_path.args) {
       return [baseToken];
